module06 almost done

+++ modules/module-6/src/app/controllers/ProductController.js
@@ -101,22 +101,28 @@ module.exports = {
             }
         }
 
-        if(req.files.length != 0){
-            const newFilesPromise =req.files.map(file =>
-                File.create({ ...file, product_id: req.body.id }))
+        const oldFiles = await Product.files(req.body.id)
+        const totalFiles = oldFiles.rows.length + req.files.length
 
-            await Promise.all(newFilesPromise)    
-        }
+        if (totalFiles <= 6) {
 
-        if(req.body.removed_files) {
-            //1,2,3,
-            const removedFiles = req.body.removed_files.split(",")//[1,2,3]
-            const lastIndex = removedFiles.length - 1
-            removedFiles.splice(lastIndex, 1)//[1,2,3]
+            if(req.body.removed_files) {
+                //1,2,3,
+                const removedFiles = req.body.removed_files.split(",")//[1,2,3]
+                const lastIndex = removedFiles.length - 1
+                removedFiles.splice(lastIndex, 1)//[1,2,3]
+    
+                const removedFilesPromise = removedFiles.map(id => File.delete(id))
+    
+                await Promise.all(removedFilesPromise)
+            }
 
-            const removedFilesPromise = removedFiles.map(id => File.delete(id))
-
-            await Promise.all(removedFilesPromise)
+            if(req.files.length != 0){
+                const newFilesPromise =req.files.map(file =>
+                    File.create({ ...file, product_id: req.body.id }))
+    
+                await Promise.all(newFilesPromise)    
+            }
         }
 
         req.body.price = req.body.price.replace(/\D/g, "")
